move dart functions to structure.rs
